scripts/requirements.txt
# python-telegram-bot==13.7
selenium
pymongo
python-dotenv
fake_useragent
webdriver_manager
colorama
psutil
Flask
pproxy
selenium-wire
selenium_stealth
undetected_chromedriver